refactor: move away from unittest entirely
